Fix read to string

+++ include/common/file_system/file_system.h
@@ -73,7 +73,7 @@ ErrnoError get_file_time_last_modification(const std::string& file_path,
 bool find_file_in_path(const std::string& file_name, std::string* out_path) WARN_UNUSED_RESULT;
 
 bool read_file_to_string(const std::string& path, std::string* contents);
-bool read_file_to_string_with_max_size(const std::string& path, std::string* contents, size_t max_size);
+bool read_file_to_string_with_max_size(const std::string& path, std::string* contents, off_t max_size);
 
 }  // namespace file_system
 }  // namespace common
